pyproject.toml
[project]
name = "nixchirp"
version = "0.1.0"
description = "Lightweight Linux-first VTuber PNGTubing app"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["vtuber", "pngtubing", "avatar", "streaming", "obs", "linux"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "PySDL2>=0.9.16",
    "pysdl2-dll>=2.0",
    "PyOpenGL>=3.1",
    "PyOpenGL-accelerate>=3.1",
    "av>=12.0",
    "Pillow>=10.0",
    "numpy>=1.24",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
input = [
    "sounddevice>=0.4",
    "alsa-midi>=0.3",
    "dbus-fast>=2.0",
]
gui = [
    "imgui-bundle>=1.0",
]
all = [
    "nixchirp[input,gui]",
]

[project.scripts]
nixchirp = "nixchirp.__main__:main"

[tool.setuptools.packages.find]
include = ["nixchirp*"]

[tool.setuptools.package-data]
"nixchirp.shaders" = ["*.vert", "*.frag"]
"nixchirp.data" = ["*.desktop", "*.xml", "icons/**/*.png"]
"nixchirp.data.examples" = ["*.toml", "*.gif", "*.apng", "*.webm"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
